Create a base class API
GithubAPI inherits from it. So would GitlabAPI.
